+++ src/test/kotlin/com/github/lucsartes/intellijprojectidentifierplugin/core/IdentifierServiceImplTest.kt
@@ -9,22 +9,22 @@ class IdentifierServiceImplTest {
 
     @Test
     fun generate_mapsMyAwesomeProjectToMAP() {
-        assertEquals("MAP", service.generate("My Awesome Project"))
+        assertEquals("MAP", service.generate("My Awesome Project", emptySet()))
     }
 
     @Test
     fun generate_handlesMultipleSpaces() {
-        assertEquals("MSH", service.generate("  many   spaces  here "))
+        assertEquals("MSH", service.generate("  many   spaces  here ", emptySet()))
     }
 
     @Test
     fun generate_handlesHyphensAndUnderscores() {
-        assertEquals("FBB", service.generate("foo-bar_baz"))
+        assertEquals("FBB", service.generate("foo-bar_baz", emptySet()))
     }
 
     @Test
     fun generate_returnsEmptyForBlankOrNoAlphanumeric() {
-        assertEquals("", service.generate("   "))
-        assertEquals("", service.generate("!!!"))
+        assertEquals("", service.generate("   ", emptySet()))
+        assertEquals("", service.generate("!!!", emptySet()))
     }
 }